putting enums where they belong
