+++ src/ingest_validation_tools/validation_utils.py
@@ -2,7 +2,6 @@ import json
 import logging
 from collections import defaultdict
 from csv import DictReader
-from enum import Enum, unique
 from pathlib import Path, PurePath
 from typing import DefaultDict, Dict, List, Optional, Union
 
@@ -12,59 +11,13 @@ from ingest_validation_tools.directory_validator import (
     DirectoryValidationErrors,
     validate_directory,
 )
+from ingest_validation_tools.enums import OtherTypes, Sample
 from ingest_validation_tools.schema_loader import (
     PreflightError,
     SchemaVersion,
     get_possible_directory_schemas,
 )
 from ingest_validation_tools.table_validator import ReportType
-
-
-@unique
-class Sample(str, Enum):
-    BLOCK = "sample-block"
-    SUSPENSION = "sample-suspension"
-    SECTION = "sample-section"
-    ORGAN = "organ"
-
-    # TODO: I believe this can be streamlined with the StrEnum class added in 3.11
-    @classmethod
-    def full_names_list(cls) -> list[str]:
-        return [sample_type.value for sample_type in cls]
-
-    @classmethod
-    def just_subtypes_list(cls) -> list[str]:
-        return [sample_type.name.lower() for sample_type in cls]
-
-    @classmethod
-    def get_key_from_val(cls, val) -> str:
-        match = [sample_type.name for sample_type in cls if sample_type.value == val]
-        if not match:
-            return ""
-        return match[0]
-
-
-@unique
-class OtherTypes(str, Enum):
-    ANTIBODIES = "antibodies"
-    CONTRIBUTORS = "contributors"
-    MURINE_SOURCE = "murine-source"
-    SAMPLE = "sample"
-    ORGAN = "organ"
-    DONOR = "donor"
-
-    @classmethod
-    def value_list(cls):
-        return [other_type.value for other_type in cls]
-
-    @classmethod
-    def get_sample_types(cls):
-        return Sample.just_subtypes_list()
-
-    @classmethod
-    def get_sample_types_full_names(cls):
-        return Sample.full_names_list()
-
 
 OTHER_TYPES_UNIQUE_FIELDS_MAP = {
     OtherTypes.ANTIBODIES: ["antibody_rrid", "antibody_name"],
